refactor(iterationEngine): extract the planning gate into a collaborator
iterationEngine.ts (1388 LOC) was not the "thin orchestrator" the docs claim:
it owned the entire pre-execution planning/readiness gate — an inline planning
provider turn, plan freshness/fingerprint logic, planning-doc writing, strict-
readiness assessment, and the ~180-line evaluatePlanningGate decision tree. The
cluster depended on only two collaborators (the CLI strategy registry and the
logger), so move it verbatim into a PlanningGate class in src/ralph/planningGate.ts
(fields named identically so the method bodies are unchanged) and have the engine
delegate via this.planningGate.evaluate(prepared).

Behaviour-preserving (iterationEngine integration + planningPass suites pass
unchanged). iterationEngine.ts drops to 956 LOC — a genuine thin orchestrator
again.
